test(remote): move testSpawnTarget into the shared helpers
runner.ci.test.ts imported it from remote.test.ts, which registers that file's
11 container tests into runner.ci's suite as well. They ran twice: the full run
reported 455 tests where 444 exist. test-utils.ts holds no top-level test(), so
importing from there costs nothing.
